requirements.txt
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==7.1.2
elasticsearch==7.9.1
fastapi==0.89.1
flake8==4.0.0
frozenlist==1.3.3
h11==0.14.0
hiredis==2.1.1
idna==3.4
mccabe==0.6.1
multidict==6.0.4
orjson==3.4.1
psycopg2-binary==2.9
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.21.0
redis==4.4.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.12.2
uvloop==0.14.0
yarl==1.8.2
elasticsearch[async]==7.9.1
